move deepspeed utils to model plugin
